feat: Add permissions to `product.controller`

+++ apps/api/src/product/product.controller.ts
@@ -3,6 +3,7 @@ import {
     UpdateProductDto,
     UpdateProductStatusDto,
     SearchProductDto,
+    User,
 } from '@fitatam/common';
 import {
     Body,
@@ -14,20 +15,42 @@ import {
     Param,
     Delete,
     ParseUUIDPipe,
+    UseGuards,
+    UnauthorizedException,
 } from '@nestjs/common';
+import { ApiTags } from '@nestjs/swagger';
+import { CookieGuard, ReqUser } from '../auth';
+import { Action, AppAbility, PoliciesGuard, UserPermissions } from '../casl';
+
 import { ProductService } from './product.service';
 
 @Controller('products')
+@ApiTags('products')
+@UseGuards(CookieGuard, PoliciesGuard)
 export class ProductController {
     constructor(private productService: ProductService) {}
 
     @Get()
-    getProducts(@Query() q: SearchProductDto) {
+    getProducts(
+        @Query() q: SearchProductDto,
+        @ReqUser() user: User,
+        @UserPermissions() perms: AppAbility
+    ) {
+        if (!perms.can(Action.Read, user)) {
+            throw new UnauthorizedException();
+        }
         return this.productService.getProducts(q);
     }
 
     @Post()
-    createProduct(@Body() dto: CreateProductDto) {
+    createProduct(
+        @Body() dto: CreateProductDto,
+        @ReqUser() user: User,
+        @UserPermissions() perms: AppAbility
+    ) {
+        if (!perms.can(Action.Create, user)) {
+            throw new UnauthorizedException();
+        }
         return this.productService.createProduct(dto);
     }
 
@@ -35,14 +58,28 @@ export class ProductController {
     @Put(':id')
     updateProduct(
         @Param('id', ParseUUIDPipe) id: string,
-        @Body() dto: UpdateProductDto
+        @Body() dto: UpdateProductDto,
+        @ReqUser() user: User,
+        @UserPermissions() perms: AppAbility
     ) {
+        if (!perms.can(Action.Update, user)) {
+            throw new UnauthorizedException();
+        }
+
         return this.productService.updateProduct(id, dto);
     }
 
     //TODO admin only
     @Delete(':id')
-    deleteProduct(@Param('id', ParseUUIDPipe) id: string) {
+    deleteProduct(
+        @Param('id', ParseUUIDPipe) id: string,
+        @ReqUser() user: User,
+        @UserPermissions() perms: AppAbility
+    ) {
+        if (!perms.can(Action.Delete, user)) {
+            throw new UnauthorizedException();
+        }
+
         return this.productService.deleteProduct(id);
     }
 
@@ -50,8 +87,14 @@ export class ProductController {
     @Post(':id/review')
     updateProductStatus(
         @Param('id', ParseUUIDPipe) id: string,
-        @Body() dto: UpdateProductStatusDto
+        @Body() dto: UpdateProductStatusDto,
+        @ReqUser() user: User,
+        @UserPermissions() perms: AppAbility
     ) {
-        return this.updateProductStatus(id, dto);
+        if (!perms.can(Action.Update, user)) {
+            throw new UnauthorizedException();
+        }
+
+        return this.productService.updateProductStatus(id, dto);
     }
 }
